Python/requirements.txt
#PyAudio==0.2.7
pyserial==2.7
requests==2.2.1